chore(watchlist-store): rename file as watchlistStore
