+++ src/store/watchlistStore.js
